Safari-611.1.10/LayoutTests/platform/gtk/http/tests/media/hls/video-controls-live-stream-expected.txt


EVENT(canplaythrough)
EVENT(play)
EXPECTED (video.duration == 'Infinity') OK
-webkit-media-text-track-container: classes: [hidden]
-webkit-media-controls-enclosure: classes: []
-webkit-media-controls-panel: classes: [paused show]
-webkit-media-controls-play-button: classes: [paused]
-webkit-media-controls-timeline: classes: []
-webkit-media-controls-current-time-display: classes: []
none: classes: [mute-box]
-webkit-media-controls-mute-button: classes: []
-webkit-media-controls-volume-slider-container: classes: [hiding]
-webkit-media-controls-volume-slider: classes: []
-webkit-media-controls-toggle-closed-captions-button: classes: [hidden]
-webkit-media-controls-fullscreen-button: classes: []
TEST(timeline == null || timeline.disabled) OK
END OF TEST
